!feat: Consistent response handling when there are no messages in the queue
Changes the response of the `/receive/pop` route if there are no messages in the queue. Instead of an empty body and the HTTP status code 204, an HTTP status code of 200 and a body of ‘{}’ (empty JSON object) will be used. Consequently, this reflects the existing behaviour when there are messages in the queue.

This will prevent the Home-Assistants Rest integration from issuing warnings, as it expects a valid JSON object in the response body.